Moves InfinityScroll outside of Homepage
