+++ src/components/InfiniteScroll/StyledInfiniteScroll.js
@@ -1,5 +1,5 @@
 import styled, { css } from 'styled-components'
-import Box from '../../Box'
+import Box from '../Box'
 
 const Container = styled(Box)`
   ${({ theme, light }) =>
